API

Trace molecule

WaterPhysics.AbstractTrace — Type

WaterPhysics uses the multiple dispatch approach to calculate the temperature and pressure dependent physical properties of water and other molecules, such as CO₂. The trace molecules and mediums are catergorized to gas or liquid subject to a general type AbstractTrace. Hierarchy of AbstractTrace:

AbstractTraceGas

AbstractTraceLiquid

abstract type AbstractTrace

WaterPhysics.AbstractTraceGas — Type

The gas can be either the target trace molecule (e.g., when computing diffusive coefficient of CO₂ in water using diffusive_coefficient) or the medium (e.g., when computing diffusive coefficient or CO₂ in air using diffusive_coefficient). Currently, WaterPhysics supports the following subtypes of AbstractTraceGas:

TraceGasAir

TraceGasCO₂

TraceGasH₂O

TraceGasO₂

abstract type AbstractTraceGas <: WaterPhysics.AbstractTrace{FT<:AbstractFloat}

WaterPhysics.AbstractTraceLiquid — Type

The liquid can be either the medium for gas (e.g., when computing diffusive coefficient of CO₂ in water using diffusive_coefficient) or the target molecule (e.g., when computing surface tension of water using surface_tension). Currently. WaterPhysics supports the following subtypes of AbstractTraceLiquid:

TraceLiquidH₂O

abstract type AbstractTraceLiquid <: WaterPhysics.AbstractTrace{FT<:AbstractFloat}

WaterPhysics.TraceGasAir — Type

Identity trace label for air.

struct TraceGasAir{FT<:AbstractFloat} <: WaterPhysics.AbstractTraceGas{FT<:AbstractFloat}

Examples

# load package before using the public functions
using WaterPhysics
# gas-phase air trace
trace = TraceGasAir{Float64}();

WaterPhysics.TraceGasCO₂ — Type

Identity trace label for gas phase CO₂.

struct TraceGasCO₂{FT<:AbstractFloat} <: WaterPhysics.AbstractTraceGas{FT<:AbstractFloat}

Fields

d_air::AbstractFloat

Diffusive coefficient in air in [m² s⁻¹]

a_298::AbstractFloat

Hydrodynamic radius of the solute in [m]

a_a::AbstractFloat

Coefficient to make temperature correction over ydrodynamic radius

d_water::AbstractFloat

Diffusive coefficient in liquid water in [m² s⁻¹]

Examples

# gas-phase CO₂ trace
trace = TraceGasCO₂{Float64}();

WaterPhysics.TraceGasH₂O — Type

Identity trace label for gas phase H₂O.

struct TraceGasH₂O{FT<:AbstractFloat} <: WaterPhysics.AbstractTraceGas{FT<:AbstractFloat}

Fields

d_air::AbstractFloat

Diffusive coefficient in air in [m² s⁻¹]

Examples

# gas-phase H₂O trace
trace = TraceGasH₂O{Float64}();

WaterPhysics.TraceGasN₂ — Type

Identity trace label for gas phase N₂.

struct TraceGasN₂{FT<:AbstractFloat} <: WaterPhysics.AbstractTraceGas{FT<:AbstractFloat}

Fields

a_298::AbstractFloat

Hydrodynamic radius of the solute in [m]

a_a::AbstractFloat

Coefficient to make temperature correction over ydrodynamic radius

d_water::AbstractFloat

Diffusive coefficient in liquid water in [m² s⁻¹]

Examples

# gas-phase N₂ trace
trace = TraceGasN₂{Float64}();

WaterPhysics.TraceGasO₂ — Type

Identity trace label for gas phase O₂.

struct TraceGasO₂{FT<:AbstractFloat} <: WaterPhysics.AbstractTraceGas{FT<:AbstractFloat}

Fields

d_air::AbstractFloat

Diffusive coefficient in air in [m² s⁻¹]

d_water::AbstractFloat

Diffusive coefficient in liquid water in [m² s⁻¹]

Examples

# gas-phase O₂ trace
trace = TraceGasO₂{Float64}();

WaterPhysics.TraceLiquidH₂O — Type

Identity trace label for liquid phase H₂O.

struct TraceLiquidH₂O{FT<:AbstractFloat} <: WaterPhysics.AbstractTraceLiquid{FT<:AbstractFloat}

Fields

γ_cor::AbstractFloat

Surface tension coefficient correction

γ_exp::AbstractFloat

Surface tension coefficient exponent

γ_k::AbstractFloat

Surface tension coefficient k in [N m⁻¹]

γ_ref::AbstractFloat

Surface tension at 298.15 K in [N m⁻¹]

γ_T_c::AbstractFloat

Surface tension coefficient T_crit in [K]

υ_A::AbstractFloat

Viscosity coefficient A in [Pa s]

υ_B::AbstractFloat

Viscosity coefficient B in [K]

υ_C::AbstractFloat

Viscosity coefficient C in [K⁻¹]

υ_D::AbstractFloat

Viscosity coefficient D in [K⁻²]

Examples

# liquid water trace
trace = TraceLiquidH₂O{Float64}();

Capillary pressure

WaterPhysics.capillary_pressure — Function

Capillary pressure of the trace liquid in a pipe is a function of surface tension (γ), pipe raduis (r), and contact angle (α):

\[P_{c} = \dfrac{2 ⋅ γ ⋅ \text{cos}(α)}{r}\]

WaterPhysics supports the following methods to compute capillary pressure:

capillary_pressure(r, T)
capillary_pressure(r, T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:338.

capillary_pressure(r, T, α)
capillary_pressure(r, T, α, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:375.

WaterPhysics.capillary_pressure — Method

When the contact angle is not given, capillary pressure is computed using

capillary_pressure(
            r::FT,
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Capillary pressure of trace liquid in [Pa], given

r Curvature radius in [m]

T Trace liquid temperature in [K]

med Medium. Optional. Default is liquid water

Examples

# capillary pressure of liquid water
p_1 = capillary_pressure(3e-6, 298.15);
p_2 = capillary_pressure(3e-6, 298.15, TraceLiquidH₂O{Float64}());

WaterPhysics.capillary_pressure — Method

When the contact angle is given, capillary pressure is computed using

capillary_pressure(
            r::FT,
            T::FT,
            α::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Capillary pressure of trace liquid in [Pa], given

r Pipe radius in [m]

T Trace liquid temperature in [K]

α Contact angle in [°]

med Medium. Optional. Default is liquid water

Examples

# capillary pressure of liquid water with contact angle
p_1 = capillary_pressure(3e-6, 298.15, 30.0);
p_2 = capillary_pressure(3e-6, 298.15, 30.0, TraceLiquidH₂O{Float64}());

Diffusive coefficient

WaterPhysics.diffusive_coefficient — Function

Diffusion of trace molecules in medium is temperature dependent, to calculate this temperature dependency, we provide a function to quick estimate this value for different trace molecules using diffusive_coefficient. The supported methods are:

diffusive_coefficient(T, mol, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:442.

diffusive_coefficient(T, mol, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:481.

WaterPhysics.diffusive_coefficient — Method

For the diffusive coefficient of gas in air, the coefficient is simply treated as a function of temperature and the reference coefficient at 298.15 K:

D = D_\text{ref} ⋅ \left( \dfrac{T}{298.15} \right)^{1.8}

This temperature correction is also available as a stand alone function: relative_diffusive_coefficient.

diffusive_coefficient(
            T::FT,
            mol::Union{TraceGasCO₂{FT}, TraceGasH₂O{FT}, TraceGasO₂{FT}},
            med::TraceGasAir{FT}
) where {FT<:AbstractFloat}

Diffusive coefficient of trace molecule in [m² s⁻¹], given

T Trace medium temperature in [K]

mol Trace molecule

med Diffusion medium (air)

Examples

# diffusive coefficient of gas-phase CO₂, H₂O, and O₂ in air
air   = TraceGasAir{Float64}();
d_CO₂ = diffusive_coefficient(298.15, TraceGasCO₂{Float64}(), air);
d_H₂O = diffusive_coefficient(298.15, TraceGasH₂O{Float64}(), air);
d_O₂  = diffusive_coefficient(298.15, TraceGasO₂{Float64}(), air);

WaterPhysics.relative_diffusive_coefficient — Function

The temperature dependency of diffusive coefficient impacts leaf gas exchange via change the maximal stomatal conductance to H₂O vapor and CO₂, given that Since the stomatal conductance should not exceed its structural limit. To account for this effect, we provide a function to calculate the diffusive coefficient relative to 25 Celcius.

relative_diffusive_coefficient(T)
relative_diffusive_coefficient(T, mol)
relative_diffusive_coefficient(T, mol, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:537.

relative_diffusive_coefficient(T, mol, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:573.

WaterPhysics.relative_diffusive_coefficient — Method

As mentioned in diffusive_coefficient, relative diffusive coefficient of target gas in gas medium is calculated as an exponential function of temperature. The shortcut method is

relative_diffusive_coefficient(
            T::FT,
            mol::AbstractTraceGas{FT} = TraceGasH₂O{FT}(),
            med::AbstractTraceGas{FT} = TraceGasAir{FT}()
) where {FT<:AbstractFloat}

Relative diffusive coefficient of trace gas in medium, given

T Water vapor temperature in [K]

mol Trace molecule. Optional, default is water vapor

med Medium. Optional, default is air

Examples

# relative diffusive coefficient of gas in air
rat_1 = relative_diffusive_coefficient(298.15);
rat_2 = relative_diffusive_coefficient(298.15, TraceGasH₂O{Float64}());
rat_3 = relative_diffusive_coefficient(298.15, TraceGasH₂O{Float64}(),
                                       TraceGasAir{Float64});

WaterPhysics.relative_diffusive_coefficient — Method

As to disolved gas diffsuin in liquid medium, the relative coefficient is also computed based on the empirical formulation from Cadogan et al. (2014).

relative_diffusive_coefficient(
            T::FT,
            mol::AbstractTraceGas{FT}
            med::AbstractTraceGas{FT}
) where {FT<:AbstractFloat}

Relative diffusive coefficient of trace gas in medium, given

T Water vapor temperature in [K]

mol Trace molecule. Optional, default is water vapor

med Medium. Optional, default is air

Examples

# relative diffusive coefficient of CO₂ and N₂ in liquid water
water = TraceLiquidH₂O{Float64}();
rat_1 = relative_diffusive_coefficient(298.15, TraceGasCO₂{Float64}(), water);
rat_2 = relative_diffusive_coefficient(298.15, TraceGasN₂{Float64}(), water);

Latent heat of evaporation

WaterPhysics.latent_heat_vapor — Function

Water evaporation from liquid phase is a key process to regulate leaf temperature, and to best represent this process. We computed the latent heat coefficient from water temperature:

latent_heat_vapor(T)
latent_heat_vapor(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:627.

WaterPhysics.latent_heat_vapor — Method

latent_heat_vapor(
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Latent heat of vaporization in [J kg⁻¹], given

T Medium temperature in [K]

med Medium. Optional. Default is liquid water

Examples

# latent heat of vapor of liquid water
λ_1 = latent_heat_vapor(298.15);
λ_2 = latent_heat_vapor(298.15, TraceLiquidH₂O{Float64}());

Surface tension of air-water interface

WaterPhysics.surface_tension — Function

When water temperature increases, the surface tension at the air-water interface decreases. Surface tension changes impacts the plant water transport via two aspects. First, if surface tension is lower, for a constant soil water content, the soil matrix potential gets less negative because the capillary pressure at the air-water interface decreases. And this is beneficial to plants. Second, the air-water interface at the pit membrane also has lower capillary pressure when temperature increases, meaning that the xylem conduits are less resistant to drought-induced air-seeded cavitation. And this is harmful for plants. Though the surface tension does not differ much with temperature change within the plant physiological active range, we account for this effect in our Land model.

surface_tension(T)
surface_tension(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:937.

WaterPhysics.surface_tension — Method

surface_tension(
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Surface tension of trace liquid in [N m⁻¹], given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

The equation used is

\[γ = 0.2358 ⋅ \left( 1 - \dfrac{T}{T_c} \right)^{1.256} ⋅ \left[ 1 - 0.625 ⋅ \left( 1 - \dfrac{T}{T_c} \right) \right]\]

See http://www.iapws.org/relguide/Surf-H2O.html

Examples

# surface tension of liquid water
γ_1 = surface_tension(298.15);
γ_2 = surface_tension(298.15, TraceLiquidH₂O{Float64}());

WaterPhysics.relative_surface_tension — Function

Relative surface tension of trace liquid relative to 298.15 K.

relative_surface_tension(T)
relative_surface_tension(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:980.

WaterPhysics.relative_surface_tension — Method

relative_surface_tension(
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Relative surface tension of trace liquid relative to 298.15 K, given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Examples

# relative surface tension of liquid water
γ_1 = relative_surface_tension(298.15);
γ_2 = relative_surface_tension(298.15, TraceLiquidH₂O{Float64}());

Vapor pressure

WaterPhysics.saturation_vapor_pressure — Function

When temperature increases, liquid water vapor pressure increases exponentially. And this correlation is accounted for using the functions below:

saturation_vapor_pressure(T)
saturation_vapor_pressure(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:736.

saturation_vapor_pressure(T, Ψ)
saturation_vapor_pressure(T, Ψ, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:780.

WaterPhysics.saturation_vapor_pressure — Method

saturation_vapor_pressure(
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Saturation vapor pressure in [Pa], given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Examples

# saturation vapor pressure of liquid water with flat surface
p_1 = saturation_vapor_pressure(298.15);
p_2 = saturation_vapor_pressure(298.15, TraceLiquidH₂O{Float64}());

WaterPhysics.saturation_vapor_pressure — Method

saturation_vapor_pressure(
            T::FT,
            Ψ::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Saturation vapor pressure in [Pa], given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface; if Ψ is given, pressure_correction is made

med Medium. Optional. Default is liquid water

Examples

# saturation vapor pressure of liquid water with non-flat surface
p_1 = saturation_vapor_pressure(298.15, -1e6);
p_2 = saturation_vapor_pressure(298.15, -1e6, TraceLiquidH₂O{Float64}());

WaterPhysics.saturation_vapor_pressure_slope — Function

First order derivative of saturation vapor pressure.

saturation_vapor_pressure_slope(T)
saturation_vapor_pressure_slope(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:829.

saturation_vapor_pressure_slope(T, Ψ)
saturation_vapor_pressure_slope(T, Ψ, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:869.

WaterPhysics.saturation_vapor_pressure_slope — Method

saturation_vapor_pressure_slope(
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

First order derivative of saturation vapor pressure in [Pa K⁻¹], given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Compute the the 1st order derivative of saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The re-arranged Clausius-Clapeyron relation

\[\frac{∂P_{sat}^{*}}{∂T} = P_{sat} ⋅ \dfrac{ LH_0 + Δc_p ⋅ (T - T_{triple})} { R_v ⋅ T^2 }\]

Examples

# slope of saturation vapor pressure of liquid water with flat surface
s_1 = saturation_vapor_pressure_slope(298.15);
s_2 = saturation_vapor_pressure_slope(298.15, TraceLiquidH₂O{Float64}());

WaterPhysics.saturation_vapor_pressure_slope — Method

saturation_vapor_pressure_slope(
            T::FT,
            Ψ::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

First order derivative of saturation vapor pressure in [Pa K⁻¹], given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface; if Ψ is given, pressure_correction is made

med Medium. Optional. Default is liquid water

Examples

# slope of saturation vapor pressure of liquid water with non-flat surface
s_1 = saturation_vapor_pressure_slope(298.15, -1e6);
s_2 = saturation_vapor_pressure_slope(298.15, -1e6, TraceLiquidH₂O{Float64}());

WaterPhysics.pressure_correction — Function

Yet, the saturation vapor pressure is not only a function of temperature, but also a function of the air-water interface curvature, known as the Kelvin equation. The package provide pressure_correction to make the correction.

pressure_correction(T, Ψ)
pressure_correction(T, Ψ, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:694.

WaterPhysics.pressure_correction — Method

pressure_correction(
            T::FT,
            Ψ::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Kelvin correction factor for saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface

med Medium. Optional. Default is liquid water

The Kelvin equation is

\[\log \left( \dfrac{P_{sat}}{P_{sat}^{*}} \right) = \dfrac{Ψ ⋅ V_{m}}{R ⋅ T}\]

Examples

# correction to saturation vapor pressure of liquid water with non-flat surface
cor_1 = pressure_correction(298.15, -1e6);
cor_2 = pressure_correction(298.15, -1e6, TraceLiquidH₂O{Float64}());

Viscosity of liquid water

WaterPhysics.viscosity — Function

When temperature increases, liquid water viscosuty decreases, meaning that the resistance for water decreases and the pressure drop per flow rate decreases. This effect is pretty significant as 1 degree increase of temperature results in about 2.4% drop in viscosity, and this is very beneficial to plant water transport. Unfortunately, to our knowledge, very few models account for this effect when modeling plant hydraulics because of the difficulty in modeling the energy budget along the flow path. We plan to have this effect accounted for in our CliMA Land model, by computing the water tempreature along the flow path and thus the viscosity change. The functions to be used are:

viscosity(T)
viscosity(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:1049.

WaterPhysics.viscosity — Method

viscosity(T::FT,
          med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Viscosity in [Pa s], given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

A = 1.856E-14 # Pa s
B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²

Examples

# viscosity of liquid water
γ_1 = viscosity(298.15);
γ_2 = viscosity(298.15, TraceLiquidH₂O{Float64}());

WaterPhysics.relative_viscosity — Function

Viscosity relative to 298.15 K

relative_viscosity(T)
relative_viscosity(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:1103.

WaterPhysics.relative_viscosity — Method

relative_viscosity(
            T::FT,
            med::TraceLiquidH₂O{FT} = TraceLiquidH₂O{FT}()
) where {FT<:AbstractFloat}

Viscosity relative to 298.15 K, given

T Liquid water temperature in [K]

liquid Optional. Default is liquid water

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²

Examples

# relative viscosity of liquid water
γ_1 = relative_viscosity(298.15);
γ_2 = relative_viscosity(298.15, TraceLiquidH₂O{Float64}());