API

WaterPhysics.relative_diffusive_coefficient — Function

relative_diffusive_coefficient(T::FT)
relative_diffusive_coefficient(T::Array{FT})

Returns the relative diffusive coefficient of water vapor in air, given

T Water vapor temperature

WaterPhysics.latent_heat_vapor — Function

latent_heat_vapor(T::FT)
latent_heat_vapor(T::Array{FT})

The specific latent heat of vaporization, given

T (Array of) temperature

WaterPhysics.saturation_vapor_pressure — Function

saturation_vapor_pressure(T::FT)
saturation_vapor_pressure(T::Array{FT})

Return the saturation vapor pressure over a plane liquid surface given

T (Array of) temperature

WaterPhysics.saturation_vapor_pressure_slope — Function

saturation_vapor_pressure_slope(T::FT)
saturation_vapor_pressure_slope(T::Array{FT})

Return the 1st order derivative of saturation vapor pressure over a plane liquid surface, given

T (Array of) temperature

Compute the the 1st order derivative of saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The re-arranged Clausius-Clapeyron relation

\[\frac{∂P_{sat}}{∂T} = P_{sat} ⋅ \dfrac{ LH_0 + Δc_p ⋅ (T - T_{triple})} { R_v ⋅ T^2 }\]

WaterPhysics.surface_tension — Function

surface_tension(T::FT)
surface_tension(T::Array{FT})

Surface tension [N m⁻¹] of water against air, given

T (Array of) temperature

The equation used is

\[γ = 0.2358 ⋅ \left( 1 - \dfrac{T}{T_c} \right)^1.256 ⋅ \left[ 1 - 0.625 ⋅ \left( 1 - \dfrac{T}{T_c} \right) \right]\]

See http://www.iapws.org/relguide/Surf-H2O.html

WaterPhysics.relative_surface_tension — Function

relative_surface_tension(T::FT)
relative_surface_tension(T::Array{FT})

Relative surface tension of water against air relative to 298.15 K, given

T (Array of) temperature

The equation used is

\[γ = 0.2358 ⋅ \left( 1 - \dfrac{T}{T_c} \right)^1.256 ⋅ \left[ 1 - 0.625 ⋅ \left( 1 - \dfrac{T}{T_c} \right) \right]\]

See http://www.iapws.org/relguide/Surf-H2O.html

A reference temperature at 298.15 K is used because the hydraulic conductances and vulnerability curves of plants are described at 298.15 K.

WaterPhysics.viscosity — Function

viscosity(T::FT)
viscosity(T::Array{FT})

Viscosity of water [Pa s], given

T (Array of) temperature

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

A = 1.856E-14 # Pa s
B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²

WaterPhysics.relative_viscosity — Function

relative_viscosity(T::FT)
relative_viscosity(T::Array{FT})

Viscosity relative to 25 degree C (298.15 K), given

T Water temperature

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²