API

Trace molecule

A multiple dispatch approach is used to calculate the temperature and pressure dependent physical properties of water and other molecules such as CO₂. The trace molecules and mediums are catergorized to Gas or Liquid subject to a general type:

WaterPhysics.AbstractTraceMolecule — Type

Identity label for trace gas or liquid. Hierarchy of AbstractTraceMolecule:

AbstractTraceGas

AbstractTraceLiquid

abstract type AbstractTraceMolecule

WaterPhysics.AbstractTraceGas — Type

Identity label for trace gas. Hierarchy of AbstractTraceGas:

TraceGasAir

TraceGasCO₂

TraceGasH₂O

abstract type AbstractTraceGas <: AbstractTraceMolecule

WaterPhysics.AbstractTraceLiquid — Type

Identity label for trace liquid. Hierarchy of AbstractTraceLiquid:

TraceLiquidH₂O

abstract type AbstractTraceLiquid <: AbstractTraceMolecule

The supported gasses include

WaterPhysics.TraceGasAir — Type

Identity label for air.

struct TraceGasAir <: AbstractTraceGas

WaterPhysics.TraceGasCO₂ — Type

Identity label for gas phase CO₂.

struct TraceGasCO₂ <: AbstractTraceGas

WaterPhysics.TraceGasH₂O — Type

Identity label for gas phase H₂O.

struct TraceGasH₂O <: AbstractTraceGas

The supported liquid includes

WaterPhysics.TraceLiquidH₂O — Type

Identity label for liquid phase H₂O.

struct TraceLiquidH₂O <: AbstractTraceLiquid

Capillary pressure

Capillary pressure of liquid water in a pipe is a function of surface tension ($\gamma$), pipe raduis ($r$), and contact angle ($\alpha$):

\[P_{c} = \dfrac{\gamma \cdot \text{cos}(\alpha)}{r}\]

WaterPhysics.capillary_pressure — Function

Capillary pressure of trace liquid.

capillary_pressure(r, T)
capillary_pressure(r, T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:179.

capillary_pressure(r, T, α)
capillary_pressure(r, T, α, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:206.

WaterPhysics.capillary_pressure — Method

capillary_pressure(
            r::FT,
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Capillary pressure of trace liquid in [Pa], given

r Curvature radius in [m]

T Trace liquid temperature in [K]

med Medium. Optional. Default is liquid water

WaterPhysics.capillary_pressure — Method

capillary_pressure(
            r::FT,
            T::FT,
            α::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Capillary pressure of trace liquid in [Pa], given

r Curvature radius in [m]

T Trace liquid temperature in [K]

α Contact angle in [°]

med Medium. Optional. Default is liquid water

Diffusive coefficient

Diffusion of trace molecules in medium is temperature dependent, to calculate this temperature dependency, we provided a function to quick estimate this value for different trace molecules

WaterPhysics.diffusive_coefficient — Function

Diffusive coefficient of trace molecule in medium

diffusive_coefficient(T, mol, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:251.

WaterPhysics.diffusive_coefficient — Method

diffusive_coefficient(
            T::FT,
            mol::TraceGasCO₂,
            med::TraceGasAir
) where {FT<:AbstractFloat}

Diffusive coefficient of trace molecule in medium (unit: m² s⁻¹), given

T Trace medium temperature in [K]

mol Trace molecule

med Diffusion medium

Diffusive coefficient of water vapor is a temperature-dependent function, this dependence impact leaf gas exchange via change the maximal stomatal conductance to water vapor as the stomatal conductance should not exceed its structural limit. To account for this effect, we provided a function to calculate the diffusive coefficient relative to 25 Celcius.

WaterPhysics.relative_diffusive_coefficient — Function

Relative diffusive coefficient of trace gas in medium

relative_diffusive_coefficient(T)
relative_diffusive_coefficient(T, mol)
relative_diffusive_coefficient(T, mol, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:288.

WaterPhysics.relative_diffusive_coefficient — Method

relative_diffusive_coefficient(
            T::FT,
            mol::AbstractTraceGas = TraceGasH₂O(),
            med::AbstractTraceGas = TraceGasAir()
) where {FT<:AbstractFloat}

Relative diffusive coefficient of trace gas in medium, given

T Water vapor temperature in [K]

mol Trace molecule. Optional, default is water vapor

med Medium. Optional, default is air

Latent heat of evaporation

Water evaporation from liquid phase is a key process to regulate leaf temperature, and to best represent this process. We computed the latent heat coefficient from water temperature:

WaterPhysics.latent_heat_vapor — Function

Latent heat of vaporization.

latent_heat_vapor(T)
latent_heat_vapor(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:330.

WaterPhysics.latent_heat_vapor — Method

latent_heat_vapor(
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Latent heat of vaporization in [J kg⁻¹], given

T Medium temperature in [K]

med Medium. Optional. Default is liquid water

Surface tension of air-water interface

When water temperature increases, the surface tension at the air-water interface decreases. Surface tension changes impacts the plant water transport via two aspects. First, if surface tension is lower, for a constant soil water content, the soil matrix potential gets less negative because the capillary pressure at the air-water interface decreases. And this is beneficial to plants. Second, the air-water interface at the pit membrane also has lower capillary pressure when temperature increases, meaning that the xylem conduits are less resistant to drought-induced air-seeded cavitation. And this is harmful for plants. Though the surface tension does not differ much with temperature change within the plant physiological active range, we account for this effect in our Land model.

WaterPhysics.surface_tension — Function

Surface tension

surface_tension(T)
surface_tension(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:583.

WaterPhysics.surface_tension — Method

surface_tension(
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Surface tension [N m⁻¹] of trace liquid, given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

The equation used is

\[γ = 0.2358 ⋅ \left( 1 - \dfrac{T}{T_c} \right)^{1.256} ⋅ \left[ 1 - 0.625 ⋅ \left( 1 - \dfrac{T}{T_c} \right) \right]\]

See http://www.iapws.org/relguide/Surf-H2O.html

WaterPhysics.relative_surface_tension — Function

Relative surface tension of trace liquid relative to 298.15 K.

relative_surface_tension(T)
relative_surface_tension(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:621.

WaterPhysics.relative_surface_tension — Method

relative_surface_tension(
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Relative surface tension of trace liquid relative to 298.15 K, given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Vapor pressure

When temperature increases, liquid water vapor pressure increases exponentially. And this correlation is accounted for using the functions below:

WaterPhysics.saturation_vapor_pressure — Function

Saturation vapor pressure.

saturation_vapor_pressure(T)
saturation_vapor_pressure(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:418.

saturation_vapor_pressure(T, Ψ)
saturation_vapor_pressure(T, Ψ, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:454.

WaterPhysics.saturation_vapor_pressure — Method

saturation_vapor_pressure(
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Saturation vapor pressure in [Pa], given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

WaterPhysics.saturation_vapor_pressure — Method

saturation_vapor_pressure(
            T::FT,
            Ψ::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface; if Ψ is given, pressure_correction is made

med Medium. Optional. Default is liquid water

WaterPhysics.saturation_vapor_pressure_slope — Function

First order derivative of saturation vapor pressure

saturation_vapor_pressure_slope(T)
saturation_vapor_pressure_slope(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:495.

saturation_vapor_pressure_slope(T, Ψ)
saturation_vapor_pressure_slope(T, Ψ, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:527.

WaterPhysics.saturation_vapor_pressure_slope — Method

saturation_vapor_pressure_slope(
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

First order derivative of saturation vapor pressure, given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Compute the the 1st order derivative of saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The re-arranged Clausius-Clapeyron relation

\[\frac{∂P_{sat}^{*}}{∂T} = P_{sat} ⋅ \dfrac{ LH_0 + Δc_p ⋅ (T - T_{triple})} { R_v ⋅ T^2 }\]

WaterPhysics.saturation_vapor_pressure_slope — Method

saturation_vapor_pressure_slope(
            T::FT,
            Ψ::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

First order derivative of saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface; if Ψ is given, pressure_correction is made

med Medium. Optional. Default is liquid water

Yet, the saturation vapor pressure is not only a function of temperature, but also a function of the air-water interface curvature, known as the Kelvin equation. The package provide pressure_correction to make the correction.

WaterPhysics.pressure_correction — Function

Kelvin correction factor for saturation vapor pressure.

pressure_correction(T, Ψ)
pressure_correction(T, Ψ, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:386.

WaterPhysics.pressure_correction — Method

pressure_correction(
            T::FT,
            Ψ::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Kelvin correction factor for saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface

med Medium. Optional. Default is liquid water

The Kelvin equation is

\[\log \left( \dfrac{P_{sat}}{P_{sat}^{*}} \right) = \dfrac{Ψ ⋅ V_{m}}{R ⋅ T}\]

Viscosity of liquid water

When temperature increases, liquid water viscosuty decreases, meaning that the resistance for water decreases and the pressure drop per flow rate decreases. This effect is pretty significant as 1 degree increase of temperature results in about 2.4% drop in viscosity, and this is very beneficial to plant water transport. Unfortunately, to our knowledge, very few models account for this effect when modeling plant hydraulics because of the difficulty in modeling the energy budget along the flow path. We plan to have this effect accounted for in our CliMA Land model, by computing the water tempreature along the flow path and thus the viscosity change. The functions to be used are:

WaterPhysics.viscosity — Function

Viscosity.

viscosity(T)
viscosity(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:680.

WaterPhysics.viscosity — Method

viscosity(T::FT,
          med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Viscosity in [Pa s], given

T Liquid water temperature in [K]

med Medium. Optional. Default is liquid water

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

A = 1.856E-14 # Pa s
B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²

WaterPhysics.relative_viscosity — Function

Viscosity relative to 298.15 K

relative_viscosity(T)
relative_viscosity(T, med)

defined at /home/runner/work/WaterPhysics.jl/WaterPhysics.jl/src/WaterPhysics.jl:729.

WaterPhysics.relative_viscosity — Method

relative_viscosity(
            T::FT,
            med::TraceLiquidH₂O = TraceLiquidH₂O()
) where {FT<:AbstractFloat}

Viscosity relative to 298.15 K, given

T Liquid water temperature in [K]

liquid Optional. Default is liquid water

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²