API

Capillary pressure

Capillary pressure of liquid water in a pipe is a function of surface tension ($\gamma$), pipe raduis ($r$), and contact angle ($\alpha$):

\[P_{c} = \dfrac{\gamma \cdot \text{cos}(\alpha)}{r}\]

WaterPhysics.capillary_pressure — Function

capillary_pressure(r::FT, T::FT) where {FT<:AbstractFloat}
capillary_pressure(r::FT, T::FT, α::FT) where {FT<:AbstractFloat}

Compute the capillary pressure in [Pa], given

r Curvature radius in [m]

T Water vapor temperature in [K]

α Contact angle in [°]

Diffusive coefficient of water vapor

Diffusive coefficient of water vapor is a temperature-dependent function, this dependence impact leaf gas exchange via change the maximal stomatal conductance to water vapor as the stomatal conductance should nto exceed its structural limit. To account for this effect, we provided a function to calculate the diffusive coefficient relative to 25 Celcius.

WaterPhysics.relative_diffusive_coefficient — Function

relative_diffusive_coefficient(T::FT) where {FT<:AbstractFloat}

Returns the relative diffusive coefficient of water vapor in air, given

T Water vapor temperature in [K]

Latent heat of evaporation

Water evaporation from liquid phase is a key process to regulate leaf temperature, and to best represent this process. We computed the latent heat coefficient from water temperature:

WaterPhysics.latent_heat_vapor — Function

latent_heat_vapor(T::FT) where {FT<:AbstractFloat}

The specific latent heat of vaporization, given

T Liquid water temperature in [K]

Surface tension of air-water interface

When water temperature increases, the surface tension at the air-water interface decreases. Surface tension changes impacts the plant water transport via two aspects. First, if surface tension is lower, for a constant soil water content, the soil matrix potential gets less negative because the capillary pressure at the air-water interface decreases. And this is beneficial to plants. Second, the air-water interface at the pit membrane also has lower capillary pressure when temperature increases, meaning that the xylem conduits are less resistant to drought-induced air-seeded cavitation. And this is harmful for plants. Though the surface tension does not differ much with temperature change within the plant physiological active range, we account for this effect in our Land model.

WaterPhysics.surface_tension — Function

surface_tension(T::FT) where {FT<:AbstractFloat}

Surface tension [N m⁻¹] of water against air, given

T Liquid water temperature in [K]

The equation used is

\[γ = 0.2358 ⋅ \left( 1 - \dfrac{T}{T_c} \right)^{1.256} ⋅ \left[ 1 - 0.625 ⋅ \left( 1 - \dfrac{T}{T_c} \right) \right]\]

See http://www.iapws.org/relguide/Surf-H2O.html

WaterPhysics.relative_surface_tension — Function

relative_surface_tension(T::FT) where {FT<:AbstractFloat}

Relative surface tension of water against air relative to 298.15 K, given

T Liquid water temperature in [K]

The equation used is

\[γ = 0.2358 ⋅ \left( 1 - \dfrac{T}{T_c} \right)^{1.256} ⋅ \left[ 1 - 0.625 ⋅ \left( 1 - \dfrac{T}{T_c} \right) \right]\]

See http://www.iapws.org/relguide/Surf-H2O.html

A reference temperature at 298.15 K is used because the hydraulic conductances and vulnerability curves of plants are described at 298.15 K.

Vapor pressure

When temperature increases, liquid water vapor pressure increases exponentially. And this correlation is accounted for using the functions below:

WaterPhysics.saturation_vapor_pressure — Function

saturation_vapor_pressure(T::FT) where {FT<:AbstractFloat}
saturation_vapor_pressure(T::FT, Ψ::FT) where {FT<:AbstractFloat}

Return the saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface; if Ψ is given, pressure_correction is made

WaterPhysics.saturation_vapor_pressure_slope — Function

saturation_vapor_pressure_slope(T::FT) where {FT<:AbstractFloat}
saturation_vapor_pressure_slope(T::FT, Ψ::FT) where {FT<:AbstractFloat}

Return the 1st order derivative of saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface; if Ψ is given, pressure_correction is made

Compute the the 1st order derivative of saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The re-arranged Clausius-Clapeyron relation

\[\frac{∂P_{sat}^{*}}{∂T} = P_{sat} ⋅ \dfrac{ LH_0 + Δc_p ⋅ (T - T_{triple})} { R_v ⋅ T^2 }\]

Yet, the saturation vapor pressure is not only a function of temperature, but also a function of the air-water interface curvature, known as the Kelvin equation. The package provide pressure_correction to make the correction.

WaterPhysics.pressure_correction — Function

pressure_correction(T::FT, Ψ::FT) where {FT<:AbstractFloat}

Make Kelvin correction to saturation vapor pressure, given

T Liquid water temperature in [K]

Ψ Liquid water pressure in [Pa], positive/negative for convex/concave interface

The Kelvin equation is

\[\log \left( \dfrac{P_{sat}}{P_{sat}^{*}} \right) = \dfrac{Ψ ⋅ V_{m}}{R ⋅ T}\]

Viscosity of liquid water

When temperature increases, liquid water viscosuty decreases, meaning that the resistance for water decreases and the pressure drop per flow rate decreases. This effect is pretty significant as 1 degree increase of temperature results in about 2.4% drop in viscosity, and this is very beneficial to plant water transport. Unfortunately, to our knowledge, very few models account for this effect when modeling plant hydraulics because of the difficulty in modeling the energy budget along the flow path. We plan to have this effect accounted for in our CliMA Land model, by computing the water tempreature along the flow path and thus the viscosity change. The functions to be used are:

WaterPhysics.viscosity — Function

viscosity(T::FT) where {FT<:AbstractFloat}

Viscosity of water in [Pa s], given

T Liquid water temperature in [K]

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

A = 1.856E-14 # Pa s
B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²

WaterPhysics.relative_viscosity — Function

relative_viscosity(T::FT) where {FT<:AbstractFloat}

Viscosity relative to 25 degree C (298.15 K), given

T Liquid water temperature in [K]

Equation used is

\[υ = A ⋅ \exp \left( \dfrac{B}{T} + C⋅T + D⋅T^2 \right)\]

and the fitting parameters are from Reid, Prausnitz, & Poling (1987), valid through 273-643 K

B = 4209      # K
C = 0.04527   # K⁻¹
D = -3.376E-5 # K⁻²